requirements.txt
streamlit
google-cloud-firestore
google-cloud-storage
firebase
Pillow
PyDrive
gsheetsdb==0.1.13
